Level Design & Polish
Added more features to the Main level such as:

- Pushing Cube
This type of obstacle pushes the player in the direction the cube is moving to when the player collides with it.

- Rotating platform
A rotating platform that makes the player follow the rotational movement of the platform.

- Sweeper
This type of obstacle is a huge, rotating gate that, like the pushing cube, forces the player to follow the direction of movement of the gate the player collides with.

Other:
Removed unused props from the scene.